requirements.txt
pillow==10.2.0
pytesseract==0.3.10
numpy==1.26.4
opencv-python==4.9.0.80 